Extract MenuBar & StatusBar out of Layout
